// src/api/generated/Schemas/user.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * invoice management system
 * OpenAPI spec version: 1.0.0
 */

/**
 * user object
 */
export interface User {
  /** Unique identifier for the user. */
  id?: number;
  /** Username of the user. */
  username: string;
  /**
     * Password of the user.
     * @nullable
     */
  password?: string | null;
  /** First name of the user. */
  firstName: string;
  /** Last name of the user. */
  lastName: string;
  /** ZIP code of the user's address. */
  zipCode: string;
  /** City of the user's address. */
  city: string;
  /** Street of the user's address. */
  street: string;
  /** Email address of the user. */
  email: string;
}